+++ test/widget_test.dart
@@ -7,8 +7,8 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:belov_dmitry_sergeevich/fnc/app.dart';
 
-import 'package:belov_dmitry_sergeevich/main.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
